requirements.txt
fastapi
uvicorn
tensorflow-cpu
opencv-python-headless
pillow
pydantic
sqlalchemy
python-multipart
requests
